refactor: consolidating functions, new module for playing a single hand
