+++ console_display.py
@@ -12,7 +12,7 @@ def display_all_hands(hands):
 def display_all_winners(winning_hands):
     number_of_winners = len(winning_hands)
     if number_of_winners == 0:
-        print('No players had one or more pairs in their hand.')
+        print('No players had any pairs in their hand.')
     elif number_of_winners == 1:
         print(f'\n{winning_hands[0]["name"]} has won with {winning_hands[0]["pairs"]} pair(s)!')
     else:
